+++ all/hex__base64/index.d.ts
@@ -1,0 +1,2 @@
+export declare function hex__base64_(hex:string):string
+