+++ all/is_entry_file/index.d.ts
@@ -1,0 +1,4 @@
+export declare function is_entry_file_(
+	url:string,
+	entry_file?:string
+):Promise<boolean>