+++ all/uuid/index.d.ts
@@ -3,3 +3,5 @@ export {
 	uuid_ as _uuid,
 	uuid_ as uuid,
 }
+export declare function short_uuid_(uuid?:string):string
+export declare function polyfill_uuid_(uuid?:string):string