+++ fs/index.js
@@ -1,1 +1,2 @@
 export * from '../all/file_exists/index.js'
+export * from '../all/is_entry_file/index.js'