+++ uuid/index.js
@@ -1,1 +1,2 @@
+export * from '../all/hex__base64/index.js'
 export * from '../all/uuid/index.js'